requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
exceptiongroup==1.2.1
frozenlist==1.4.1
h11==0.14.0
idna==3.7
multidict==6.0.5
outcome==1.3.0.post0
packaging==24.0
pycparser==2.22
PySocks==1.7.1
python-box==7.1.1
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
selenium==4.20.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
webdriver-controller==0.0.4
webdriver-manager==4.0.1
wsproto==1.2.0
xmltodict==0.13.0
yarl==1.9.4
